Renamed tamper detection bundle
