+++ kura/org.eclipse.kura.rest.tamper.detection.provider/META-INF/MANIFEST.MF
@@ -1,0 +1,22 @@
+Manifest-Version: 1.0
+Bundle-ManifestVersion: 2
+Bundle-Name: org.eclipse.kura.rest.tamper.detection.provider
+Bundle-SymbolicName: org.eclipse.kura.rest.tamper.detection.provider;singleton:=true
+Bundle-Version: 1.0.0.qualifier
+Bundle-Vendor: Eclipse Kura
+Require-Capability: osgi.ee;filter:="(&(osgi.ee=JavaSE)(version=17))"
+Bundle-ClassPath: .
+Bundle-ActivationPolicy: lazy
+Service-Component: OSGI-INF/*.xml
+Import-Package: com.google.gson;version="2.7.0",
+ jakarta.annotation.security;version="2.1.1",
+ jakarta.ws.rs;version="3.1.0",
+ jakarta.ws.rs.core;version="3.1.0",
+ org.eclipse.kura;version="[1.0,2.0)",
+ org.eclipse.kura.cloudconnection.message;version="[1.0,2.0)",
+ org.eclipse.kura.cloudconnection.request;version="[1.0,2.0)",
+ org.eclipse.kura.message;version="[1.4,2.0)",
+ org.eclipse.kura.security.tamper.detection;version="[1.0,2.0)",
+ org.eclipse.kura.type;version="[1.1,2.0)",
+ org.osgi.service.useradmin;version="1.1.0",
+ org.slf4j;version="1.7.25"
