+++ kura/org.eclipse.kura.rest.tamper.detection.provider/about_files/epl-v20.html
