+++ kura/org.eclipse.kura.rest.tamper.detection.provider/build.properties
@@ -1,0 +1,23 @@
+#
+#  Copyright (c) 2011, 2020 Eurotech and/or its affiliates and others
+#
+#  This program and the accompanying materials are made
+#  available under the terms of the Eclipse Public License 2.0
+#  which is available at https://www.eclipse.org/legal/epl-2.0/
+#
+#  SPDX-License-Identifier: EPL-2.0
+#
+#  Contributors:
+#   Eurotech
+#
+output.. = target/classes/
+bin.includes = META-INF/,\
+               .,\
+               OSGI-INF/,\
+               about.html,\
+               about_files/,\
+               OSGI-INF/org.eclipse.kura.rest.tamper.detection.provider.TamperDetectionRestService.xml,\
+               OSGI-INF/org.eclipse.kura.rest.tamper.detection.provider.TamperDetectionRequestHandler.xml
+src.includes = about.html,\
+               about_files/
+source.. = src/main/java
