+++ kura/org.eclipse.kura.rest.tamper.detection.provider/src/main/java/org/eclipse/kura/rest/tamper/detection/api/TamperDetectionServiceInfo.java
@@ -1,0 +1,26 @@
+/*******************************************************************************
+ * Copyright (c) 2021 Eurotech and/or its affiliates and others
+ * 
+ * This program and the accompanying materials are made
+ * available under the terms of the Eclipse Public License 2.0
+ * which is available at https://www.eclipse.org/legal/epl-2.0/
+ * 
+ * SPDX-License-Identifier: EPL-2.0
+ * 
+ * Contributors:
+ *  Eurotech
+ ******************************************************************************/
+package org.eclipse.kura.rest.tamper.detection.api;
+
+public class TamperDetectionServiceInfo {
+
+    @SuppressWarnings("unused")
+    private final String pid;
+    @SuppressWarnings("unused")
+    private final String displayName;
+
+    public TamperDetectionServiceInfo(String pid, String displayName) {
+        this.pid = pid;
+        this.displayName = displayName;
+    }
+}
