+++ kura/test/org.eclipse.kura.rest.tamper.detection.provider.test/META-INF/MANIFEST.MF
@@ -1,0 +1,36 @@
+Manifest-Version: 1.0
+Bundle-ManifestVersion: 2
+Bundle-Name: org.eclipse.kura.rest.tamper.detection.provider.test
+Bundle-SymbolicName: org.eclipse.kura.rest.tamper.detection.provider.test;singleton:=true
+Bundle-Version: 6.0.0.qualifier
+Bundle-Vendor: Eclipse Kura
+Require-Capability: osgi.ee;filter:="(&(osgi.ee=JavaSE)(version=17))"
+Bundle-ActivationPolicy: lazy
+Require-Bundle: org.eclipse.kura.rest.tamper.detection.provider;bundle-version="1.0.0",
+ org.eclipse.kura.http.server.manager;bundle-version="1.1.0",
+ org.eclipse.kura.broker.artemis.core;bundle-version="1.2.0",
+ org.eclipse.kura.broker.artemis.simple.mqtt;bundle-version="1.1.0"
+Import-Package: com.google.gson;version="2.7.0",
+ com.google.gson.reflect;version="2.7.0",
+ org.apache.commons.io;version="2.4.0",
+ org.eclipse.kura;version="1.6.0",
+ org.eclipse.kura.cloud;version="1.1.0",
+ org.eclipse.kura.configuration;version="1.2.0",
+ org.eclipse.kura.core.testutil.requesthandler;version="1.0.0",
+ org.eclipse.kura.core.util;version="1.2.0",
+ org.eclipse.kura.crypto;version="1.3.0",
+ org.eclipse.kura.data;version="1.1.2",
+ org.eclipse.kura.data.transport.listener;version="1.0.1",
+ org.eclipse.kura.marshalling;version="1.0.0",
+ org.eclipse.kura.message;version="1.4.0",
+ org.eclipse.kura.security.tamper.detection;version="1.0.0",
+ org.eclipse.kura.type;version="1.1.0",
+ org.eclipse.kura.util.wire.test;version="1.1.0",
+ org.junit,
+ org.junit.runner;version="[4.12.0,5.0.0)",
+ org.junit.runners;version="[4.12.0,5.0.0)",
+ org.mockito;version="[4.0.0,5.0.0)",
+ org.mockito.stubbing;version="[4.0.0,5.0.0)",
+ org.osgi.framework;version="1.10.0",
+ org.osgi.util.tracker;version="1.5.2",
+ org.slf4j;version="1.7.25"
