+++ kura/test/org.eclipse.kura.rest.tamper.detection.provider.test/build.properties
